+++ packages/runtime/src/index.ts
@@ -1,4 +1,6 @@
 export * from "./runtime-dsl-parser";
 export * from "./dependency-graph";
+export * from "./function-registry";
+export * from "./rule-engine";
 export * from "./template-resolver";
 export * from "./types";